feat: Fix logout functionality in userSlice and redirect to login screen
The userSlice in the redux/slices directory was updated to include a new logoutUser reducer. This reducer sets the user state to null when called, effectively logging out the user. Additionally, the userProfile/UserProfile.js file was modified to redirect the user to the login screen upon logout. This change improves the logout functionality and ensures a smooth user experience.